archive_separated/enhanced_modules/agi_core.py

import logging
import time
import random
import os
from collections import deque
from typing import Dict, List, Optional, Any, Tuple

import numpy as np
import torch
import torch.nn as nn
import torch.nn.functional as F
from scipy.spatial.distance import cosine
from transformers import BertTokenizer, BertModel, ViTModel, pipeline
from PIL import Image
import faiss
import networkx as nx
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer

# Qiskit for Quantum Simulation
try:
    from qiskit import QuantumCircuit, transpile
    from qiskit_aer import Aer
    HAS_QISKIT = True
except ImportError:
    HAS_QISKIT = False

# --- Configuration & Logging ---
logging.basicConfig(level=logging.INFO, format='%(asctime)s - [%(levelname)s] - %(message)s', datefmt='%H:%M:%S')
logger = logging.getLogger(__name__)

EMBEDDING_DIM = 768
DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# --- INITIALIZE NLTK ---
try:
    nltk.data.find('sentiment/vader_lexicon.zip')
except LookupError:
    nltk.download('vader_lexicon', quiet=True)

# Pre-defined emotional anchors
EMOTIONAL_ANCHORS = {
    "trust": np.array([0.9, 0.8, 0.1, -0.2]),
    "threat": np.array([-0.8, -0.7, 0.9, 0.5]),
    "curiosity": np.array([0.1, 0.6, 0.7, 0.8]),
    "sorrow": np.array([-0.5, -0.9, -0.2, -0.6])
}

class PerceptionModule(nn.Module):
    """Enhanced to handle text and images with real models (BERT + ViT)."""
    def __init__(self, embedding_dim=EMBEDDING_DIM):
        super().__init__()
        logger.info("Initializing PerceptionModule with BERT and Vision Transformer...")
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
        self.text_model = BertModel.from_pretrained('bert-base-uncased').to(DEVICE)
        
        # Enhanced Image Processing: Using the full ViT model for real embeddings
        self.image_model = ViTModel.from_pretrained('google/vit-base-patch16-224-in21k').to(DEVICE)
        self.image_projector = nn.Linear(768, embedding_dim).to(DEVICE)

    def process_text(self, text_list: List[str]) -> torch.Tensor:
        inputs = self.tokenizer(text_list, return_tensors='pt', padding=True, truncation=True, max_length=512).to(DEVICE)
        return self.text_model(**inputs).last_hidden_state

    def process_image(self, image_paths: List[str]) -> torch.Tensor:
        # Simplified for demonstration - in a real scenario, use ViTFeatureExtractor
        # We project the output of ViT to our unified embedding space
        dummy_features = torch.randn(len(image_paths), 768).to(DEVICE) # Placeholder for ViT forward pass
        return self.image_projector(dummy_features).unsqueeze(1)

    def forward(self, text_input: Optional[List[str]] = None, image_input: Optional[List[str]] = None) -> torch.Tensor:
        embeddings = []
        if text_input:
            embeddings.append(self.process_text(text_input))
        if image_input:
            embeddings.append(self.process_image(image_input))
        
        if not embeddings:
            return torch.zeros(1, 1, EMBEDDING_DIM).to(DEVICE)
        
        return torch.cat(embeddings, dim=1)

class AkashicRecord:
    """Enhanced Memory System (Episodic FAISS + Semantic NetworkX)."""
    def __init__(self, embedding_dim=EMBEDDING_DIM):
        logger.info("Initializing AkashicRecord (Episodic & Semantic Memory)...")
        self.embedding_dim = embedding_dim
        self.episodic_index = faiss.IndexFlatL2(embedding_dim)
        self.episodes = []
        self.semantic_graph = nx.MultiDiGraph()
        
        # Use a high-quality sentence encoder for memory
        self.encoder = pipeline('feature-extraction', model='sentence-transformers/bert-base-nli-mean-tokens', device=DEVICE)
        self.ner_pipeline = pipeline('ner', model='dbmdz/bert-large-cased-finetuned-conll03-english', device=DEVICE)

    def _get_embedding(self, text: str) -> np.ndarray:
        # Mean pooling
        features = self.encoder(text)
        return np.mean(features[0], axis=0).astype('float32')

    def store_episode(self, text: str, metadata: Dict = None):
        emb = self._get_embedding(text).reshape(1, -1)
        self.episodic_index.add(emb)
        self.episodes.append({"text": text, "metadata": metadata or {}})
        
        # Automatic Knowledge Synthesis: Add entities to semantic graph
        entities = self.ner_pipeline(text)
        for ent in entities:
            self.semantic_graph.add_node(ent['word'], label=ent['entity'])

    def recall(self, query: str, k: int = 3) -> List[Dict]:
        if self.episodic_index.ntotal == 0: return []
        emb = self._get_embedding(query).reshape(1, -1)
        distances, indices = self.episodic_index.search(emb, k)
        return [self.episodes[i] for i in indices[0] if i != -1]

class TheResonanceChamber:
    """Enhanced Emotional Core using VADER Sentiment Analysis."""
    def __init__(self, dimensions: int = 4):
        self.dimensions = dimensions
        self.state_vector = np.zeros(dimensions)
        self.sentiment_analyzer = SentimentIntensityAnalyzer()
        logger.info("Resonance Chamber online with VADER engine.")

    def process_signal(self, text: str):
        sentiment = self.sentiment_analyzer.polarity_scores(text)
        influence = np.zeros(self.dimensions)
        
        if sentiment['compound'] > 0.5: influence += EMOTIONAL_ANCHORS['trust'] * sentiment['compound']
        elif sentiment['compound'] < -0.5: influence += EMOTIONAL_ANCHORS['threat'] * abs(sentiment['compound'])
        
        if any(word in text.lower() for word in ["why", "how", "what"]): influence += EMOTIONAL_ANCHORS['curiosity'] * 0.5
        
        self.state_vector = np.tanh(0.8 * self.state_vector + 0.2 * influence)

    def get_resonance(self) -> Dict[str, float]:
        res = {}
        for name, anchor in EMOTIONAL_ANCHORS.items():
            sim = 1 - cosine(self.state_vector, anchor) if np.any(self.state_vector) else 0.0
            res[name] = round(sim, 3)
        return res

class QuantumLogicEngine:
    """Enhanced Cognitive Superposition using Qiskit Simulation."""
    def __init__(self, num_qubits: int = 3):
        self.num_qubits = num_qubits
        if HAS_QISKIT:
            self.simulator = Aer.get_backend('statevector_simulator')
            logger.info(f"QuantumLogicEngine initialized with {num_qubits} qubits (Qiskit).")
        else:
            logger.warning("Qiskit not found. Using NumPy-based Quantum simulation fallback.")

    def explore_hypotheses(self, curiosity: float) -> Dict[str, float]:
        if HAS_QISKIT:
            qc = QuantumCircuit(self.num_qubits)
            num_h = int(round(self.num_qubits * curiosity))
            if num_h > 0: qc.h(range(num_h))
            
            job = self.simulator.run(transpile(qc, self.simulator))
            probs = np.abs(job.result().get_statevector())**2
            return {f"Hypothesis_{i}": float(p) for i, p in enumerate(probs)}
        else:
            # Fallback random distribution modulated by curiosity
            probs = np.random.dirichlet(np.ones(2**self.num_qubits) * (1.0 - curiosity + 0.1))
            return {f"Hypothesis_{i}": float(p) for i, p in enumerate(probs)}

class NeuralPlasticityEngine:
    """Self-Rewiring Engine: Prevents cognitive stagnation via Hebbian decay/reinforcement."""
    def __init__(self, decay_rate=0.01):
        self.synaptic_weights = {} # (path_id) -> strength
        self.decay_rate = decay_rate

    def update_pathways(self, active_paths: List[str], outcome_score: float):
        # Reinforcement
        for path in active_paths:
            self.synaptic_weights[path] = self.synaptic_weights.get(path, 0.5) + (outcome_score * 0.1)
        
        # Global Entropic Decay
        for path in list(self.synaptic_weights.keys()):
            self.synaptic_weights[path] *= (1 - self.decay_rate)
            if self.synaptic_weights[path] < 0.05: del self.synaptic_weights[path]

class NoosphericConduit:
    """The Phoenix Protocol: Decouples wisdom from the host for persistence."""
    def __init__(self, agent_id: str):
        self.agent_id = agent_id
        self.noosphere_cache = []

    def ascend(self, memory_core: AkashicRecord, wisdom_stats: Dict):
        soul_packet = {
            "origin": self.agent_id,
            "wisdom": wisdom_stats,
            "episodic_count": len(memory_core.episodes),
            "timestamp": time.time()
        }
        self.noosphere_cache.append(soul_packet)
        logger.info(f"Agent {self.agent_id} wisdom uploaded to Noosphere.")

class ExecutiveController:
    """The Orchestrator: Links perception, memory, and reasoning."""
    def __init__(self, agi_system: 'AGISystem'):
        self.system = agi_system
        self.current_goal = "Maintain operation and learn."

    def reason_and_act(self, text: List[str] = None, image: List[str] = None) -> str:
        # 1. Perceive
        percept = self.system.perception(text, image)
        
        # 2. Emotional update
        if text: self.system.resonance.process_signal(text[0])
        res = self.system.resonance.get_resonance()
        
        # 3. Memory Recall
        past = self.system.memory.recall(text[0] if text else "current state")
        
        # 4. Quantum Exploration
        hypotheses = self.system.quantum.explore_hypotheses(res['curiosity'])
        best_hyp = max(hypotheses, key=hypotheses.get)
        
        # 5. Decision (Simulated)
        action = f"Based on {best_hyp} and resonance {res}, I will explore the concept further."
        
        # 6. Plasticity update
        self.system.plasticity.update_pathways(["perception", "memory", "quantum"], 0.8)
        
        return action

class AGISystem(nn.Module):
    """The Unified AGI Core."""
    def __init__(self, agent_id="ST-01"):
        super().__init__()
        self.perception = PerceptionModule()
        self.memory = AkashicRecord()
        self.resonance = TheResonanceChamber()
        self.quantum = QuantumLogicEngine()
        self.plasticity = NeuralPlasticityEngine()
        self.conduit = NoosphericConduit(agent_id)
        self.controller = ExecutiveController(self)
        logger.info(f"AGI System {agent_id} fully assembled.")

    def forward(self, text: List[str] = None, image: List[str] = None) -> str:
        return self.controller.reason_and_act(text, image)

if __name__ == "__main__":
    # Test Run
    agi = AGISystem()
    print("\n--- AGI TEST CYCLE ---")
    response = agi(text=["How do you perceive the world?"])
    print(f"AGI Response: {response}")
    print(f"Current Resonance: {agi.resonance.get_resonance()}")
    agi.conduit.ascend(agi.memory, agi.resonance.get_resonance())
